postgresql: resolve dead lock situation for some tests
Function execute-with-new-tx (cf. its docstring) makes database
'introspection' functions (database-list-tables, database-list-views
etc) always execute in a fresh transaction.

If a CREATE VIEW was not yet committed in the original transaction, the
listing of the views in a new TX would hang indefinitely, waiting for a
lock in the original TX.

Did NOT remote the function completely, as I didn't make a thorough
clean up.